WinKnownPaths

WinKnownPaths.path

WinKnownPaths.path — Method

folderpath(folderid::UUID) -> AbstractString

Windows only function to retrieve the full path of a known folder identified by the folder's id
specified in the module `FOLDERID`. The FOLDERID module contains the folderid constants available for query.

Examples

julia> import WinKnownPaths: FOLDERID

julia> WinKnownPaths.path(FOLDERID.System)
"C:\WINDOWS\system32"

julia> WinKnownPaths.path(FOLDERID.Fonts)
"C:\WINDOWS\Fonts"